backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0
sqlalchemy==2.0.30
alembic==1.13.1
pymysql==1.1.1
cryptography==42.0.8
pydantic==2.7.4
pydantic-settings==2.3.4
PyJWT==2.8.0
bcrypt==4.1.3
python-multipart==0.0.9
cloudinary==1.40.0
python-dotenv==1.0.1
email-validator==2.1.1
httpx==0.27.0
